third_party/blink/web_tests/external/wpt/png/exif-chunk-expected.txt

This is a testharness.js-based test.
[FAIL] test pixel values of a rotated PNG
  assert_approx_equals: expected 0 +/- 2 but got 255
Harness: the test ran to completion.
